moved-user-fuctions-outofmain
